move ethrex l2 init command to ethrex l2
