+++ cmd/ethrex/l2/mod.rs
@@ -4,6 +4,7 @@ pub mod command;
 pub mod options;
 
 pub use command::Command;
+pub use initializers::init_l2;
 pub use options::{
     BlockProducerOptions, CommitterOptions, EthOptions, Options as L2Options,
     ProofCoordinatorOptions, SequencerOptions, WatcherOptions,
